feat(console): abstract away node_options exp. await
